refactor: split shop services
